requirements.txt
absl-py==0.13.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
docopt==0.6.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.17
grpcio==1.39.0
idna==3.2
importlib-metadata==4.8.1
ipython==7.27.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.2
networkx==2.6.2
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
pandas==1.3.2
parso==0.8.2
pdf2image==1.16.0
pickleshare==0.7.5
Pillow==8.3.1
pipreqs==0.4.10
prompt-toolkit==3.0.20
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pygraphviz==1.7
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
torch==1.9.0
torchsummary==1.5.1
torchvision==0.10.0
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wincertstore==0.2
yarg==0.1.9
zipp==3.5.0
